+++ src/main/java/tesis/controllers/UsersController.java
@@ -9,20 +9,38 @@ import tesis.services.impl.PocService;
 import tesis.services.impl.UsersService;
 
 @RestController  // o @Controller
-@RequestMapping("/usersxxx/create_user")
+@RequestMapping("/usersXX")
 public class UsersController {
     @Autowired
     UsersService usersService;
 
-    @PostMapping("/response")
+    @PostMapping("/create_user_hc")
+    @ResponseBody
+    //public String createUser(@RequestParam("user_id") Long userId, @RequestParam("role") String role ) {
+    public String createUserHC(
+            @RequestBody Users newuser)
+    {
+        usersService.create_user_hc();
+        return ("Usuario Creado (hard coded @ service)");
+    }
+
+    @PostMapping("/create_user")
     @ResponseBody
     //public String createUser(@RequestParam("user_id") Long userId, @RequestParam("role") String role ) {
     public String createUser(
-           @RequestBody Users newuser)
+            @RequestBody Users newuser)
     {
-        return ("Usuario Creado - fake");
+        usersService.create_user(newuser);
+        return ("Usuario Creado (4real)");
     }
 
+    @PostMapping("/create_user_fake")
+    public Users createUserFake(@RequestBody Users newuserfake) {
+        System.out.println("entró al post");
+        //return pocService.getDTO(123);
+        //usersService.createUser();
+        return usersService.getUserHC(1234L);
+    }
     /*
     public ResponseTransfer postResponseController(
             @RequestBody LoginForm loginForm) {
